refactor(callbacks): extract US Grid tab helpers to _callbacks_us_grid.py
Step 2 of issue #87. Pulls the US Grid tab's data-collection + render
helpers (cards / map / polygons) out of ``callbacks.py`` into
``components/_callbacks_us_grid.py``. Also promotes
``_latest_real_demand`` to ``_callbacks_shared.py`` since it's
cross-tab (Overview hero chart + US Grid card-grid collector).

callbacks.py: 7,051 → 6,389 lines (-662, ~9.4%). Combined with PR #98,
total: 7,167 → 6,389 (-778 lines, -11%).

## Moved (720 lines) → _callbacks_us_grid.py

- ``_collect_us_grid_region_data`` — Redis fan-out for the 51-BA card grid
- ``_build_us_grid_title`` — page title + V3.ζ hidden-region annotation
- ``_is_real_positive`` — NaN guard used across US Grid arithmetic
- ``_build_us_grid_metrics_items`` — 4-up MetricsBar with stress cap
- ``_build_us_grid_sparkline`` — per-card 24h sparkline SVG
- ``_build_interchange_chip`` — V3.α net interchange chip
- ``_build_us_grid_region_card`` — single region card (with V3.η imports chip)
- ``_build_us_grid_map`` — Plotly ``scatter_geo`` BA centroid view
- ``_load_ba_polygons`` + ``_BA_POLYGONS_CACHE`` — geojson loader + cache
- ``_build_us_grid_choropleth`` — Plotly ``Choropleth`` polygon view (PR #79/#80)

## Promoted to _callbacks_shared.py

- ``_latest_real_demand`` — NaN-aware "latest valid demand" helper,
  used by both Overview hero chart and US Grid collector. Moving here
  avoids a circular import callbacks → us_grid → callbacks.

## Public-import surface preserved

``components/callbacks.py`` re-imports every moved symbol via explicit
named imports + ``# noqa: F401`` re-export markers. All 1,688 tests
pass with zero changes to ``app.py``. The US Grid test files (4 of
them) needed module-path updates for ``patch()`` / ``monkeypatch``
calls — the patches now target the function's *new* module
namespace, not the old callbacks.py one.

## Test-path updates (mechanical, no behavior change)

- ``components.callbacks.redis_get`` → ``components._callbacks_us_grid.redis_get``
- ``components.callbacks.REGION_NAMES`` → ``components._callbacks_us_grid.REGION_NAMES``
- ``components.callbacks.REGION_CAPACITY_MW`` → ``components._callbacks_us_grid.REGION_CAPACITY_MW``
- ``components.callbacks._load_ba_polygons`` → ``components._callbacks_us_grid._load_ba_polygons``

This is the standard "patch where it's used, not where it's defined"
fix that follows any module-relocation refactor — the function reads
its imports from its OWN module namespace, so patches need to target
that namespace.

## Verification

- All 1,688 tests pass
- ``ruff check`` + ``ruff format --check`` clean
- ``app.py``'s ``from components.callbacks import _BACKTEST_CACHE, _MODEL_CACHE,
  _PREDICTION_CACHE`` (line 179) still resolves correctly
- Polygon view rendering preserved (regression tests pass)
